move api outside of operations
