+++ .graphqlrc.yml
@@ -14,7 +14,7 @@ extensions:
     emitLegacyCommonJSImports: false
     overwrite: true
     generates:
-      ./libs/data-layer/operations/api.ts:
+      ./libs/data-layer/api.ts:
         plugins:
           - add:
               content: *HEADER
