+++ libs/data-layer/operations/queries/FamilyListQuery/useFamilyListQuery.tsx
@@ -3,8 +3,8 @@ import * as Apollo from '@apollo/client';
 import type {
   ApiFamilyListQuery,
   ApiFamilyListQueryVariables,
-} from '../../api';
-import { ApiFamilyDocument } from '../../api';
+} from '../../../api';
+import { ApiFamilyDocument } from '../../../api';
 import type { InformationStatus, SeverityStatus } from '../FamilyQuery';
 import { toInformationStatus, toSeverityStatus } from '../FamilyQuery';
